@@ -1,8 +1,13 @@
 module github.com/alexboor/lbx-telebot
 
-go 1.20
+go 1.21
 
-require gopkg.in/telebot.v3 v3.1.3
+require (
+	github.com/jackc/pgx/v4 v4.18.1
+	github.com/jdkato/prose/v2 v2.0.0
+	golang.org/x/exp v0.0.0-20230801115018-d63ba01acd4b
+	gopkg.in/telebot.v3 v3.1.3
+)
 
 require (
 	github.com/deckarep/golang-set v1.7.1 // indirect
@@ -13,12 +18,9 @@ require (
 	github.com/jackc/pgproto3/v2 v2.3.2 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
 	github.com/jackc/pgtype v1.14.0 // indirect
-	github.com/jackc/pgx/v4 v4.18.1 // indirect
 	github.com/jackc/puddle v1.3.0 // indirect
-	github.com/jdkato/prose/v2 v2.0.0 // indirect
 	github.com/mingrammer/commonregex v1.0.1 // indirect
 	golang.org/x/crypto v0.6.0 // indirect
-	golang.org/x/exp v0.0.0-20230801115018-d63ba01acd4b // indirect
 	golang.org/x/text v0.7.0 // indirect
 	gonum.org/v1/gonum v0.7.0 // indirect
 	gopkg.in/neurosnap/sentences.v1 v1.0.6 // indirect
